Move SWIM to a separate package
